+++ classes/local/pricemodifier/modifiers/taxes.php
@@ -99,12 +99,12 @@ abstract class taxes extends modifier_base {
                     $itemisnet = get_config('local_shopping_cart', 'itempriceisnet');
                     $iseuropean = vatnrchecker::is_european($countrycode);
                     $isowncountry = vatnrchecker::is_own_country($countrycode);
-                    $hasvarnr = $cartstore->has_vatnr_data();
+                    $hasvatnumber = $cartstore->has_vatnr_data();
                     if ($itemisnet) {
                         $netprice = $items[$key]['price']; // Price is now considered a net price.
                         if (
                             $iseuropean &&
-                            $hasvarnr &&
+                            $hasvatnumber &&
                             !$isowncountry
                         ) {
                             // EU reverse charge applies. No taxes added to net.
@@ -128,7 +128,7 @@ abstract class taxes extends modifier_base {
                         $netprice = round($items[$key]['price'] / (1 + $taxpercent), 2);
                         if (
                             $iseuropean &&
-                            $hasvarnr &&
+                            $hasvatnumber &&
                             !$isowncountry
                         ) {
                             $grossprice = $netprice;